requirements.txt
audioread==3.0.1
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.6
cycler==0.12.1
Cython==3.0.10
data==0.4
decorator==5.1.1
distro==1.9.0
docstring-to-markdown==0.15
docutils==0.21.2
fonttools==4.51.0
funcsigs==1.0.2
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
jedi==0.19.1
joblib==1.4.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
matplotlib==3.9.0
mido==1.3.2
msgpack==1.0.8
numba==0.59.1
numpy==1.26.4
packaging==23.2
paramiko==3.4.0
parso==0.8.4
pathlib==1.0.1
PGen==0.2.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.1
pretty_midi==0.2.10
PyAudio==0.2.14
pycparser==2.22
pygame-ce==2.4.1
pygame_gui==0.6.10
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pypiwin32==223
PySoundFile==0.9.0.post1
python-dateutil==2.9.0.post0
python-i18n==0.3.9
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
pywin32==306
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.5.0
tk==0.1.0
typing_extensions==4.12.0
ujson==5.9.0
urllib3==2.2.1
zipp==3.18.2
